update home screen widgets and order screen widgets
