+++ lib/widgets/home_screen_widgets/home_product_sales.dart
@@ -1,0 +1,115 @@
+import 'package:flutter/material.dart';
+import 'package:syncfusion_flutter_charts/charts.dart';
+
+class ProductSales extends StatelessWidget {
+  const ProductSales({
+    Key? key,
+  }) : super(key: key);
+
+  @override
+  Widget build(BuildContext context) {
+    return Row(
+      children: [
+        Flexible(
+          child: Card(
+            elevation: 5,
+            shadowColor: Colors.blueGrey,
+            child: Column(
+              crossAxisAlignment: CrossAxisAlignment.start,
+              children: const <Widget>[
+                Padding(
+                  padding: EdgeInsets.only(top: 10.0, left: 10.0),
+                  child: Text(
+                    "Product Sales",
+                    style: TextStyle(
+                        fontSize: 20,
+                        color: Colors.green,
+                        fontWeight: FontWeight.w400),
+                  ),
+                ),
+                ChartDateButton(),
+                SalesChart()
+              ],
+            ),
+          ),
+        ),
+      ],
+    );
+  }
+}
+
+class ChartDateButton extends StatelessWidget {
+  const ChartDateButton({
+    Key? key,
+  }) : super(key: key);
+
+  @override
+  Widget build(BuildContext context) {
+    return Row(
+      mainAxisAlignment: MainAxisAlignment.end,
+      crossAxisAlignment: CrossAxisAlignment.center,
+      children: [
+        Padding(
+          padding: const EdgeInsets.only(right: 8.0),
+          child: ElevatedButton(onPressed: () {}, child: Text("Day")),
+        ),
+        Padding(
+          padding: const EdgeInsets.only(right: 8.0),
+          child: ElevatedButton(onPressed: () {}, child: Text("Month")),
+        ),
+        Padding(
+          padding: const EdgeInsets.only(right: 8.0),
+          child: ElevatedButton(onPressed: () {}, child: Text("Year")),
+        ),
+      ],
+    );
+  }
+}
+
+class SalesChart extends StatefulWidget {
+  const SalesChart({
+    Key? key,
+  }) : super(key: key);
+
+  @override
+  State<SalesChart> createState() => _SalesChartState();
+}
+
+class _SalesChartState extends State<SalesChart> {
+  late TooltipBehavior _tooltipBehavior;
+
+  @override
+  void initState() {
+    _tooltipBehavior = TooltipBehavior(enable: true);
+    super.initState();
+  }
+
+  @override
+  Widget build(BuildContext context) {
+    return SfCartesianChart(
+        primaryXAxis: CategoryAxis(),
+        legend: Legend(isVisible: false),
+        tooltipBehavior: _tooltipBehavior,
+        series: <LineSeries<SalesData, String>>[
+          LineSeries<SalesData, String>(
+            dataSource: <SalesData>[
+              SalesData('Jan', 35),
+              SalesData('Feb', 28),
+              SalesData('Mar', 34),
+              SalesData('Apr', 32),
+              SalesData('May', 40)
+            ],
+            xValueMapper: (SalesData sales, _) => sales.year,
+            yValueMapper: (SalesData sales, _) => sales.sales,
+            // Enable data label
+            dataLabelSettings: DataLabelSettings(isVisible: true),
+          ),
+        ]);
+  }
+}
+
+class SalesData {
+  SalesData(this.year, this.sales);
+  final String year;
+  final double sales;
+}
